restructures the code to fix cycle dependencies
