+++ src/pkg/retention/policy/action/index/index.go
@@ -16,9 +16,8 @@ package index
 
 import (
 	"github.com/goharbor/harbor/src/lib/errors"
-	"github.com/goharbor/harbor/src/pkg/immutable/match/rule"
 	"github.com/goharbor/harbor/src/pkg/retention/policy/action"
-	policyindex "github.com/goharbor/harbor/src/pkg/retention/policy/rule/index"
+	"github.com/goharbor/harbor/src/pkg/retention/policy/action/performer"
 	"sync"
 )
 
@@ -27,7 +26,7 @@ var index sync.Map
 
 func init() {
 	// Register retain action
-	Register(action.Retain, action.NewRetainAction(rule.NewRuleMatcher(policyindex.Get)))
+	Register(action.Retain, performer.NewRetainAction)
 }
 
 // Register the performer with the corresponding action
