+++ src/pkg/retention/policy/alg/or/processor_test.go
@@ -16,8 +16,7 @@ package or
 
 import (
 	"errors"
-	rule2 "github.com/goharbor/harbor/src/pkg/immutable/match/rule"
-	policyindex "github.com/goharbor/harbor/src/pkg/retention/policy/rule/index"
+	"github.com/goharbor/harbor/src/pkg/retention/policy/action/performer"
 	"testing"
 	"time"
 
@@ -31,7 +30,6 @@ import (
 	"github.com/goharbor/harbor/src/lib/selector/selectors/doublestar"
 	"github.com/goharbor/harbor/src/lib/selector/selectors/label"
 	"github.com/goharbor/harbor/src/pkg/retention/dep"
-	"github.com/goharbor/harbor/src/pkg/retention/policy/action"
 	"github.com/goharbor/harbor/src/pkg/retention/policy/alg"
 	"github.com/goharbor/harbor/src/pkg/retention/policy/rule"
 	"github.com/goharbor/harbor/src/pkg/retention/policy/rule/always"
@@ -89,8 +87,7 @@ func (suite *ProcessorTestSuite) TearDownSuite() {
 // TestProcess tests process method
 func (suite *ProcessorTestSuite) TestProcess() {
 
-	factory := action.NewRetainAction(rule2.NewRuleMatcher(policyindex.Get))
-	perf := factory(suite.all, false)
+	perf := performer.NewRetainAction(suite.all, false)
 	params := make([]*alg.Parameter, 0)
 	lastxParams := make(map[string]rule.Parameter)
 	lastxParams[lastx.ParameterX] = 10
@@ -131,8 +128,7 @@ func (suite *ProcessorTestSuite) TestProcess() {
 
 // TestProcess2 ...
 func (suite *ProcessorTestSuite) TestProcess2() {
-	factory := action.NewRetainAction(rule2.NewRuleMatcher(policyindex.Get))
-	perf := factory(suite.all, false)
+	perf := performer.NewRetainAction(suite.all, false)
 
 	params := make([]*alg.Parameter, 0)
 	alwaysParams := make(map[string]rule.Parameter)
