+++ src/pkg/retention/policy/builder_test.go
@@ -15,8 +15,7 @@
 package policy
 
 import (
-	rule2 "github.com/goharbor/harbor/src/pkg/immutable/match/rule"
-	policyindex "github.com/goharbor/harbor/src/pkg/retention/policy/rule/index"
+	"github.com/goharbor/harbor/src/pkg/retention/policy/action/performer"
 	"testing"
 	"time"
 
@@ -90,7 +89,7 @@ func (suite *TestBuilderSuite) SetupSuite() {
 		doublestar.NSExcludes,
 	}, doublestar.New)
 	index.Register(label.Kind, []string{label.With, label.Without}, label.New)
-	index3.Register(action.Retain, action.NewRetainAction(rule2.NewRuleMatcher(policyindex.Get)))
+	index3.Register(action.Retain, performer.NewRetainAction)
 
 	suite.oldClient = dep.DefaultClient
 	dep.DefaultClient = &fakeRetentionClient{}
