+++ src/pkg/retention/policy/rule/dayspl/evaluator.go
@@ -16,12 +16,12 @@ package dayspl
 
 import (
 	"fmt"
+	"github.com/goharbor/harbor/src/pkg/retention/policy/action"
 	"time"
 
 	"github.com/goharbor/harbor/src/common/utils"
 	"github.com/goharbor/harbor/src/lib/log"
 	"github.com/goharbor/harbor/src/lib/selector"
-	"github.com/goharbor/harbor/src/pkg/retention/policy/action"
 	"github.com/goharbor/harbor/src/pkg/retention/policy/rule"
 )
