+++ src/pkg/retention/policy/rule/index/index.go
@@ -15,10 +15,10 @@
 package index
 
 import (
+	"github.com/goharbor/harbor/src/pkg/retention/policy/action"
 	"sync"
 
 	"github.com/goharbor/harbor/src/lib/errors"
-	"github.com/goharbor/harbor/src/pkg/retention/policy/action"
 	"github.com/goharbor/harbor/src/pkg/retention/policy/rule"
 	"github.com/goharbor/harbor/src/pkg/retention/policy/rule/always"
 	"github.com/goharbor/harbor/src/pkg/retention/policy/rule/dayspl"
