+++ src/pkg/retention/policy/rule/latestps/evaluator.go
@@ -16,13 +16,13 @@ package latestps
 
 import (
 	"fmt"
+	"github.com/goharbor/harbor/src/pkg/retention/policy/action"
 	"math"
 	"sort"
 
 	"github.com/goharbor/harbor/src/common/utils"
 	"github.com/goharbor/harbor/src/lib/log"
 	"github.com/goharbor/harbor/src/lib/selector"
-	"github.com/goharbor/harbor/src/pkg/retention/policy/action"
 	"github.com/goharbor/harbor/src/pkg/retention/policy/rule"
 )
